Set node value to -1 when ISY reports it as unknown (blank string)
There is a very distinct difference between _blank_ values and 0 values as reported by the ISY. Blank values are used when the ISY does not know the status of a node, such as in cases where a battery-powered sensor has not reported a state since the ISY last booted up. The current implementation would cause leak sensors, for example, to report as WET (the dry node would be 0). This change makes the value be -1 to hopefully reduce the magnitude of breaking change that this could be, while still allowing consuming applications to treat UNKNOWN values as different than OFF