Fix Programs throwing errors when ISY in 24-hour Clock Mode
The time string returned for Last Start and Last Finish times returned with the initial `/rest/programs' call varies depending on if the ISY is in 24-hr mode or not.

Originally reported [here](https://community.home-assistant.io/t/insteon-thermostat-2441th-for-isy/97821/53?u=shbatm).

This fix is a "patch" and just checks to see if there are 2 spaces in the time string. I have a more advanced (read: overkill) method staged for v2 which actually includes the `rest/time` endpoint and provides some additional ISY clock/date/location info.