Modify Group Status Method to show an all_on property (#72)
The Group (scene) status method has been updated to use list comprehension to determine the group state. It will now also return an aux_property for "all_on" -- true if all devices in the scene are on, false if only some of the devices are on.
Flip order for group off update for smooth transition.