[V2] Use structured dict-based class for Aux_Properties (#77)
* Correct UOM for Ramp Rate in Aux Properties

* Get formatted value from correct location in event stream XML

Formatted value is passed as a `fmtAct` tag in the event stream, not as an attribute for the value like if calling `/rest/nodes/ADDRESS`

* Use structured dict-based class for Aux_Properties

Include migration of a few more constants used in other integrations to string literals.

Move and rename EventResult class to NodeProperty class. Moved to helpers to prevent circular imports.
Cleanup

Use NodeProperty for Aux_Props

* Correct extra quotes in strings. Use new NodeProperty formatting.

* Correct logic for adding ramp rates to aux_properties from nodes.xml

Else/Elif Logic was malformed in the ramp rate check when processing nodes.xml. Ramp rate would be translated, but never added to aux_props.

Logic was correct when picking up the ramp rate in the event stream.

* iSort and Black Formatting

* Corrections to dict inherited class.

* Simplify Logging Call