Finally fix automicus/PyISY#11 (#86)
Manually check if single nodes are a parent for other nodes.

This was not possible using the existing code because the way the root assignment and get_by_id of the Nodes class is set up, you can never return a Nodes class (group of nodes) for a single Node ID, regardless of if it has children or not.  So all_lower_nodes, children, and has_children always returned empty for a node.

Also updated the REPR function to print Group Members.