[V2] Use Instance instead of Class Variables To Support Multiple ISYs (#92)
Eliminate the use of Class variables in favor of Instance Variables. Using Class Variables for the navigator classes causes an issue when multiple ISYs are in use.  Only one instance of PyISY is loaded, so it attempt to share the same lists for all instances of the ISY class. When multiple ISYs are connected in the same program, this causes errors.