[V2] Missing constant, query command, get_by_name fix, and Update README (#94)
* [V2] Use Instance instead of Class Variables To Support Multiple ISYs

Eliminate the use of Class variables in favor of Instance Variables. Using Class Variables for the navigator classes causes an issue when multiple ISYs are in use.  Only one instance of PyISY is loaded, so it attempt to share the same lists for all instances of the ISY class. When multiple ISYs are connected in the same program, this causes errors.

* Add missing constant. Update README.

* Expose the ISY Query Function for All Nodes and Specific Addresses

Adds the isy.query() function to run a full or device-specific query.

* [V2] Fix get_by_name functions in nodes, programs, and add to variables

* Actually fix automicus/PyISY#50 - Don't update all vars if using auto.

* Rename parameter